refactor(*): remove unused code and files
