--- app/src/main/res/xml/settings.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen
-	xmlns:android="http://schemas.android.com/apk/res/android"
->
-
-  <PreferenceCategory
-    android:title="Agent Infos"
-  >
-  <Preference
-       android:key="device_id"
-       android:title="Device ID"
-       android:summary=""
-       android:selectable="false"
-       android:enabled="true"
-      />
-  </PreferenceCategory>
-	<PreferenceCategory android:title="@string/agent_settings">
-		<CheckBoxPreference
-			android:key="boot"
-			android:persistent="true"
-			android:defaultValue="false" android:summaryOn="@string/phone_boot" android:summaryOff="@string/manual_launch" android:title="@string/boot_options"/>
-			
-			
-	</PreferenceCategory>
-	<PreferenceCategory
-		android:title="Server Settings"
-
-	>
-		<org.flyve.inventory.agent.UrlPreference
-			android:key="url" android:title="Inventory Server" android:persistent="true" android:defaultValue="http://"/>
-
-	
-<EditTextPreference android:dialogMessage="@string/tag" android:persistent="true" android:enabled="true" android:selectable="true" android:dialogTitle="@string/tag" android:title="@string/tag" android:key="tag"></EditTextPreference>
-</PreferenceCategory><PreferenceCategory android:title="@string/authentication_credentials"><EditTextPreference android:dialogMessage="@string/login" android:enabled="true" android:persistent="true" android:title="@string/login" android:key="login"></EditTextPreference>
-<EditTextPreference android:dialogMessage="@string/password" android:persistent="true" android:title="@string/password" android:key="password" android:enabled="true"></EditTextPreference>
-</PreferenceCategory>
-</PreferenceScreen>
